Restruture tests. Add one assembler test.
